refactoring: User 관련 4-layered architecture 적용
