+++ src/main/java/com/onepage/coupong/business/user/dto/request/EmailCertificationRequestDto.java
@@ -1,5 +1,6 @@
-package com.onepage.coupong.user.dto.request;
+package com.onepage.coupong.business.user.dto.request;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,12 +9,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class SignInRequestDto {
+public class EmailCertificationRequestDto {
 
     @NotBlank
     private String username;
 
     @NotBlank
-    private String password;
-
+    @Email
+    private String email;
 }
