+++ src/main/java/com/onepage/coupong/business/user/dto/request/IdCheckRequestDto.java
@@ -1,6 +1,5 @@
-package com.onepage.coupong.user.dto.request;
+package com.onepage.coupong.business.user.dto.request;
 
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,12 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class EmailCertificationRequestDto {
+public class IdCheckRequestDto {
 
     @NotBlank
     private String username;
-
-    @NotBlank
-    @Email
-    private String email;
 }
