+++ src/main/java/com/onepage/coupong/business/user/dto/response/SignInResponseDto.java
@@ -1,6 +1,6 @@
-package com.onepage.coupong.user.dto.response;
+package com.onepage.coupong.business.user.dto.response;
 
-import com.onepage.coupong.user.domain.enums.UserRole;
+import com.onepage.coupong.jpa.user.enums.UserRole;
 import com.onepage.coupong.global.response.ResponseCode;
 import com.onepage.coupong.global.response.ResponseMessage;
 import lombok.Getter;
