+++ src/main/java/com/onepage/coupong/infrastructure/mail/MailServiceImpl.java
@@ -1,11 +1,11 @@
 package com.onepage.coupong.infrastructure.mail;
 
-import com.onepage.coupong.user.dto.request.CheckEmailCertificationRequestDto;
-import com.onepage.coupong.user.dto.request.EmailCertificationRequestDto;
-import com.onepage.coupong.user.dto.response.ResponseDto;
-import com.onepage.coupong.user.dto.response.CheckEmailCertificationResponseDto;
-import com.onepage.coupong.user.dto.response.EmailCertificationResponseDto;
-import com.onepage.coupong.user.domain.Certification;
+import com.onepage.coupong.business.user.dto.request.CheckEmailCertificationRequestDto;
+import com.onepage.coupong.business.user.dto.request.EmailCertificationRequestDto;
+import com.onepage.coupong.business.user.dto.response.ResponseDto;
+import com.onepage.coupong.business.user.dto.response.CheckEmailCertificationResponseDto;
+import com.onepage.coupong.business.user.dto.response.EmailCertificationResponseDto;
+import com.onepage.coupong.jpa.user.Certification;
 import jakarta.mail.Message;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
