+++ src/main/java/com/onepage/coupong/jpa/coupon/CouponAttemptLog.java
@@ -1,6 +1,6 @@
 package com.onepage.coupong.jpa.coupon;
 
-import com.onepage.coupong.user.domain.User;
+import com.onepage.coupong.jpa.user.User;
 import com.onepage.coupong.jpa.coupon.enums.CouponCategory;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
