+++ src/main/java/com/onepage/coupong/jpa/user/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.onepage.coupong.user.domain.enums;
+package com.onepage.coupong.jpa.user.enums;
 
 public enum UserRole {
     ROLE_USER, ROLE_ADMIN
